move wasm not sync types into wasm_not_sync module
